Clarify thc ipc recv/send, add return statuses to some functions.
    -- thc_ipc_send ==> thc_ipc_send_request
    -- thc_ipc_recv ==> thc_ipc_recv_response
    -- thc_ipc_recv_response ignores requests and messages that
       are neither requests or responses (i.e., it only handles
       responses). See updated header documentation.
    -- awe_mapper_create_id and awe_mapper_get_id no longer
       BUG on error conditions; they return a non-zero exit status
       instead.
    -- THCYieldToId and THCYieldToIdAndSave return non-zero if
       the id does not match an awe.

Not built or tested yet. Updated examples. (In the context switch
example, I just BUG on in there, so you have the same semantics
as before.)

+++ src/include/thc_ipc_types.h
@@ -14,8 +14,8 @@ typedef enum {
 
 struct predicate_payload
 {
-    unsigned long expected_msg_id;
-    unsigned long actual_msg_id;
+    uint32_t expected_msg_id;
+    uint32_t actual_msg_id;
     msg_type_t msg_type;
 };
